+++ cmd/logger_unix.go
@@ -1,0 +1,8 @@
+//go:build !windows
+// +build !windows
+
+package cmd
+
+func toZapLogPath(p string) string {
+	return p
+}